+++ pmd-core/src/test/java/net/sourceforge/pmd/cpd/CSVRendererTest.java
@@ -13,7 +13,6 @@ import net.sourceforge.pmd.PMD;
 import org.junit.Test;
 
 public class CSVRendererTest {
-
     @Test
     public void testLineCountPerFile() {
         Renderer renderer = new CSVRenderer(true);
@@ -32,16 +31,29 @@ public class CSVRendererTest {
         assertEquals(expectedReport, report);
     }
 
+    @Test
+    public void testFilenameEscapes() {
+        Renderer renderer = new CSVRenderer();
+        List<Match> list = new ArrayList<>();
+        String codeFragment = "code\nfragment";
+        Mark mark1 = createMark("public", "/var,with,commas/Foo.java", 48, 10, codeFragment);
+        Mark mark2 = createMark("stuff", "/var,with,commas/Bar.java", 73, 20, codeFragment);
+        Match match = new Match(75, mark1, mark2);
+        list.add(match);
+
+        String report = renderer.render(list.iterator());
+        String expectedReport =
+                "lines,tokens,occurrences" + PMD.EOL +
+                "10,75,2,48,\"/var,with,commas/Foo.java\",73,\"/var,with,commas/Bar.java\"" + PMD.EOL;
+        assertEquals(expectedReport, report);
+    }
+
     private Mark createMark(String image, String tokenSrcID, int beginLine, int lineCount, String code) {
         Mark result = new Mark(new TokenEntry(image, tokenSrcID, beginLine));
 
         result.setLineCount(lineCount);
         result.setSoureCodeSlice(code);
         return result;
-    }
-
-    public static junit.framework.Test suite() {
-        return new junit.framework.JUnit4TestAdapter(CSVRendererTest.class);
     }
 }